requirements.txt
PyYAML
fastapi==0.111.0
uvicorn==0.30.1
tensorflow
pillow
numpy>=1.24.0
pydantic>=2.0.0
gunicorn==22.0.0
requests>=2.25.0
